move static files to model properties data
