+++ src/marqo_commons/model_registry/model_properties_data/clip_properties.py
@@ -4,7 +4,7 @@ CLIP Model Properties
 This file defines properties for CLIP models. It is intended to be used in conjunction with the model registry
 and should not be used in isolation.
 """
-from marqo_commons.model_registry.model_properties_object import ModelProperties, VectorNumericType, Modality, ModelType
+from marqo_commons.model_registry.model_properties_data.model_properties_object import ModelProperties, VectorNumericType, Modality, ModelType
 from typing import Dict, List
 from marqo_commons.model_registry.utils import convert_model_properties_to_dict
