+++ src/marqo_commons/model_registry/model_registry.py
@@ -1,16 +1,16 @@
 from typing import Dict
 import json
 
-from marqo_commons.model_registry.sbert_properties import _get_sbert_properties
-from marqo_commons.model_registry.sbert_onnx_properties import _get_sbert_onnx_properties
-from marqo_commons.model_registry.clip_properties import _get_clip_properties
-from marqo_commons.model_registry.open_clip_properties import _get_open_clip_properties
-from marqo_commons.model_registry.multilingual_clip_properties import _get_multilingual_clip_properties
-from marqo_commons.model_registry.test_properties import _get_sbert_test_properties
-from marqo_commons.model_registry.hf_properties import _get_hf_properties
-from marqo_commons.model_registry.fp16_clip_properties import _get_fp16_clip_properties
-from marqo_commons.model_registry.onnx_clip_properties import _get_onnx_clip_properties
-from marqo_commons.model_registry.random_properties import _get_random_properties
+from marqo_commons.model_registry.model_properties_data.sbert_properties import _get_sbert_properties
+from marqo_commons.model_registry.model_properties_data.sbert_onnx_properties import _get_sbert_onnx_properties
+from marqo_commons.model_registry.model_properties_data.clip_properties import _get_clip_properties
+from marqo_commons.model_registry.model_properties_data.open_clip_properties import _get_open_clip_properties
+from marqo_commons.model_registry.model_properties_data.multilingual_clip_properties import _get_multilingual_clip_properties
+from marqo_commons.model_registry.model_properties_data.test_properties import _get_sbert_test_properties
+from marqo_commons.model_registry.model_properties_data.hf_properties import _get_hf_properties
+from marqo_commons.model_registry.model_properties_data.fp16_clip_properties import _get_fp16_clip_properties
+from marqo_commons.model_registry.model_properties_data.onnx_clip_properties import _get_onnx_clip_properties
+from marqo_commons.model_registry.model_properties_data.random_properties import _get_random_properties
 
 
 # we need to keep track of the embed dim and model load functions/classes
