+++ src/marqo_commons/shared_utils/constants.py
@@ -1,5 +1,6 @@
 MODEL_NAME_SIZE_MAPPING = {
     # This mapping represents the default sizes for the precise model names.
+    # Sizes are in gigabytes (GB)
     "vit-l-14" : 1.5,
     "vit-g" : 5,
     "vit-h" : 5,
