+++ tests/test_model_properties.py
@@ -1,17 +1,18 @@
 import json
 from unittest import TestCase
 
-from marqo_commons.model_registry.clip_properties import _get_clip_properties
-from marqo_commons.model_registry.fp16_clip_properties import _get_fp16_clip_properties
-from marqo_commons.model_registry.hf_properties import _get_hf_properties
+from marqo_commons.model_registry.model_properties_data.clip_properties import _get_clip_properties
+from marqo_commons.model_registry.model_properties_data.fp16_clip_properties import _get_fp16_clip_properties
+from marqo_commons.model_registry.model_properties_data.hf_properties import _get_hf_properties
+from marqo_commons.model_registry.model_properties_data.multilingual_clip_properties import _get_multilingual_clip_properties
+from marqo_commons.model_registry.model_properties_data.onnx_clip_properties import _get_onnx_clip_properties
+from marqo_commons.model_registry.model_properties_data.open_clip_properties import _get_open_clip_properties
+from marqo_commons.model_registry.model_properties_data.random_properties import _get_random_properties
+from marqo_commons.model_registry.model_properties_data.sbert_onnx_properties import _get_sbert_onnx_properties
+from marqo_commons.model_registry.model_properties_data.sbert_properties import _get_sbert_properties
+from marqo_commons.model_registry.model_properties_data.test_properties import _get_sbert_test_properties
+
 from marqo_commons.model_registry.model_registry import get_model_properties_dict, get_model_properties_json
-from marqo_commons.model_registry.multilingual_clip_properties import _get_multilingual_clip_properties
-from marqo_commons.model_registry.onnx_clip_properties import _get_onnx_clip_properties
-from marqo_commons.model_registry.open_clip_properties import _get_open_clip_properties
-from marqo_commons.model_registry.random_properties import _get_random_properties
-from marqo_commons.model_registry.sbert_onnx_properties import _get_sbert_onnx_properties
-from marqo_commons.model_registry.sbert_properties import _get_sbert_properties
-from marqo_commons.model_registry.test_properties import _get_sbert_test_properties
 from marqo_commons.shared_utils.enums import ModelType
